Add dark mode compatibility, change to tsx
