+++ src/components/Services/Service.tsx
@@ -1,0 +1,199 @@
+const Service = () => {
+  return (
+    <div className="flex flex-wrap gap-0 dark:text-white">
+      <a href="#" className="xl:w-1/3 md:w-1/2 p-6 cursor-pointer">
+        <div className="border border-gray-300 p-6 rounded-lg cursor-pointer">
+          <div className="w-10 h-10 inline-flex items-center justify-center rounded-full bg-indigo-100 text-indigo-500 mb-4">
+            <svg
+              fill="none"
+              stroke="currentColor"
+              strokeLinecap="round"
+              strokeLinejoin="round"
+              strokeWidth="2"
+              className="w-6 h-6"
+              viewBox="0 0 24 24"
+            >
+              <path d="M22 12h-4l-3 9L9 3l-3 9H2"></path>
+            </svg>
+          </div>
+          <h2 className="text-lg font-medium title-font mb-2">IPO Checker</h2>
+          <p className="leading-relaxed text-base">
+            Lorem Ipsum is simply dummy text of the printing and typesetting
+            industry.
+          </p>
+        </div>
+      </a>
+      <a href="#" className="xl:w-1/3 md:w-1/2 p-6 ">
+        <div className="border border-gray-300 p-6 rounded-lg cursor-pointer">
+          <div className="w-10 h-10 inline-flex items-center justify-center rounded-full bg-indigo-100 text-indigo-500 mb-4">
+            <svg
+              fill="none"
+              stroke="currentColor"
+              strokeLinecap="round"
+              strokeLinejoin="round"
+              strokeWidth="2"
+              className="w-6 h-6"
+              viewBox="0 0 24 24"
+            >
+              <circle cx="6" cy="6" r="3"></circle>
+              <circle cx="6" cy="18" r="3"></circle>
+              <path d="M20 4L8.12 15.88M14.47 14.48L20 20M8.12 8.12L12 12"></path>
+            </svg>
+          </div>
+          <h2 className="text-lg font-medium title-font mb-2">
+            IPO/FPO Updates
+          </h2>
+          <p className="leading-relaxed text-base">
+            Lorem Ipsum is simply dummy text of the printing and typesetting
+            industry.
+          </p>
+        </div>
+      </a>
+      <a href="#" className="xl:w-1/3 md:w-1/2 p-6 cursor-pointer">
+        <div className="border border-gray-300 p-6 rounded-lg">
+          <div className="w-10 h-10 inline-flex items-center justify-center rounded-full bg-indigo-100 text-indigo-500 mb-4">
+            <svg
+              fill="none"
+              stroke="currentColor"
+              strokeLinecap="round"
+              strokeLinejoin="round"
+              strokeWidth="2"
+              className="w-6 h-6"
+              viewBox="0 0 24 24"
+            >
+              <path d="M20 21v-2a4 4 0 00-4-4H8a4 4 0 00-4 4v2"></path>
+              <circle cx="12" cy="7" r="4"></circle>
+            </svg>
+          </div>
+          <h2 className="text-lg font-medium title-font mb-2">
+            Trading Signals
+          </h2>
+          <p className="leading-relaxed text-base">
+            Lorem Ipsum is simply dummy text of the printing and typesetting
+            industry.
+          </p>
+        </div>
+      </a>
+      <a href="#" className="xl:w-1/3 md:w-1/2 p-6 cursor-pointer">
+        <div className="border border-gray-300 p-6 rounded-lg">
+          <div className="w-10 h-10 inline-flex items-center justify-center rounded-full bg-indigo-100 text-indigo-500 mb-4">
+            <svg
+              fill="none"
+              stroke="currentColor"
+              strokeLinecap="round"
+              strokeLinejoin="round"
+              strokeWidth="2"
+              className="w-6 h-6"
+              viewBox="0 0 24 24"
+            >
+              <path d="M4 15s1-1 4-1 5 2 8 2 4-1 4-1V3s-1 1-4 1-5-2-8-2-4 1-4 1zM4 22v-7"></path>
+            </svg>
+          </div>
+          <h2 className="text-lg font-medium title-font mb-2">
+            Portfolio Tracker
+          </h2>
+          <p className="leading-relaxed text-base">
+            Lorem Ipsum is simply dummy text of the printing and typesetting
+            industry.
+          </p>
+        </div>
+      </a>
+      <a href="#" className="xl:w-1/3 md:w-1/2 p-6 cursor-pointer">
+        <div className="border border-gray-300 p-6 rounded-lg">
+          <div className="w-10 h-10 inline-flex items-center justify-center rounded-full bg-indigo-100 text-indigo-500 mb-4">
+            <svg
+              fill="none"
+              stroke="currentColor"
+              strokeLinecap="round"
+              strokeLinejoin="round"
+              strokeWidth="2"
+              className="w-6 h-6"
+              viewBox="0 0 24 24"
+            >
+              <path d="M21 12.79A9 9 0 1111.21 3 7 7 0 0021 12.79z"></path>
+            </svg>
+          </div>
+          <h2 className="text-lg font-medium title-font mb-2">Charts</h2>
+          <p className="leading-relaxed text-base">
+            Lorem Ipsum is simply dummy text of the printing and typesetting
+            industry.
+          </p>
+        </div>
+      </a>
+      <a href="#" className="xl:w-1/3 md:w-1/2 p-6 cursor-pointer">
+        <div className="border border-gray-300 p-6 rounded-lg">
+          <div className="w-10 h-10 inline-flex items-center justify-center rounded-full bg-indigo-100 text-indigo-500 mb-4">
+            <svg
+              fill="none"
+              stroke="currentColor"
+              strokeLinecap="round"
+              strokeLinejoin="round"
+              strokeWidth="2"
+              className="w-6 h-6"
+              viewBox="0 0 24 24"
+            >
+              <path d="M12 22s8-4 8-10V5l-8-3-8 3v7c0 6 8 10 8 10z"></path>
+            </svg>
+          </div>
+          <h2 className="text-lg font-medium title-font mb-2">
+            Company Analyser
+          </h2>
+          <p className="leading-relaxed text-base">
+            Lorem Ipsum is simply dummy text of the printing and typesetting
+            industry.
+          </p>
+        </div>
+      </a>
+      <a href="#" className="xl:w-1/3 md:w-1/2 p-6 cursor-pointer">
+        <div className="border border-gray-300 p-6 rounded-lg">
+          <div className="w-10 h-10 inline-flex items-center justify-center rounded-full bg-indigo-100 text-indigo-500 mb-4">
+            <svg
+              fill="none"
+              stroke="currentColor"
+              strokeLinecap="round"
+              strokeLinejoin="round"
+              strokeWidth="2"
+              className="w-6 h-6"
+              viewBox="0 0 24 24"
+            >
+              <path d="M12 22s8-4 8-10V5l-8-3-8 3v7c0 6 8 10 8 10z"></path>
+            </svg>
+          </div>
+          <h2 className="text-lg font-medium title-font mb-2">
+            Compare Company
+          </h2>
+          <p className="leading-relaxed text-base">
+            Lorem Ipsum is simply dummy text of the printing and typesetting
+            industry.
+          </p>
+        </div>
+      </a>
+      <a href="#" className="xl:w-1/3 md:w-1/2 p-6 cursor-pointer">
+        <div className="border border-gray-300 p-6 rounded-lg">
+          <div className="w-10 h-10 inline-flex items-center justify-center rounded-full bg-indigo-100 text-indigo-500 mb-4">
+            <svg
+              fill="none"
+              stroke="currentColor"
+              strokeLinecap="round"
+              strokeLinejoin="round"
+              strokeWidth="2"
+              className="w-6 h-6"
+              viewBox="0 0 24 24"
+            >
+              <path d="M12 22s8-4 8-10V5l-8-3-8 3v7c0 6 8 10 8 10z"></path>
+            </svg>
+          </div>
+          <h2 className="text-lg font-medium title-font mb-2">
+            Share Calculator
+          </h2>
+          <p className="leading-relaxed text-base">
+            Lorem Ipsum is simply dummy text of the printing and typesetting
+            industry.
+          </p>
+        </div>
+      </a>
+    </div>
+  );
+};
+
+export default Service;
